requirements.txt
# --- Web Framework & Server ---
# Required for the local dashboard and background lifespan management
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic-settings>=2.2.1

# --- Hardware & System Monitoring ---
# Required for detecting CPU/RAM and tracking idle states
psutil>=5.9.8

# Optional but highly recommended: Detects NVIDIA GPUs for AI workloads
GPUtil>=1.4.0

# --- Containerization ---
# Required for the script to safely talk to the local Docker Desktop daemon
docker>=7.0.0

# --- Network & File Communication ---
# Required for making async HTTP requests to your Master Server
httpx>=0.27.0

# Required for saving large AI datasets to the hard drive asynchronously
# without freezing the background engine
aiofiles>=23.2.1